+++ oci/data_safe_data_safe_private_endpoint_resource.go
@@ -1,0 +1,527 @@
+// Copyright (c) 2017, 2019, Oracle and/or its affiliates. All rights reserved.
+
+package oci
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"strings"
+	"time"
+
+	"github.com/hashicorp/terraform/helper/resource"
+	"github.com/hashicorp/terraform/helper/schema"
+
+	oci_common "github.com/oracle/oci-go-sdk/common"
+	oci_data_safe "github.com/oracle/oci-go-sdk/datasafe"
+)
+
+func init() {
+	RegisterResource("oci_data_safe_data_safe_private_endpoint", DataSafeDataSafePrivateEndpointResource())
+}
+
+func DataSafeDataSafePrivateEndpointResource() *schema.Resource {
+	return &schema.Resource{
+		Importer: &schema.ResourceImporter{
+			State: schema.ImportStatePassthrough,
+		},
+		Timeouts: DefaultTimeout,
+		Create:   createDataSafeDataSafePrivateEndpoint,
+		Read:     readDataSafeDataSafePrivateEndpoint,
+		Update:   updateDataSafeDataSafePrivateEndpoint,
+		Delete:   deleteDataSafeDataSafePrivateEndpoint,
+		Schema: map[string]*schema.Schema{
+			// Required
+			"compartment_id": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"display_name": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"subnet_id": {
+				Type:     schema.TypeString,
+				Required: true,
+				ForceNew: true,
+			},
+			"vcn_id": {
+				Type:     schema.TypeString,
+				Required: true,
+				ForceNew: true,
+			},
+
+			// Optional
+			"defined_tags": {
+				Type:             schema.TypeMap,
+				Optional:         true,
+				Computed:         true,
+				DiffSuppressFunc: definedTagsDiffSuppressFunction,
+				Elem:             schema.TypeString,
+			},
+			"description": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Computed: true,
+			},
+			"freeform_tags": {
+				Type:     schema.TypeMap,
+				Optional: true,
+				Computed: true,
+				Elem:     schema.TypeString,
+			},
+			"nsg_ids": {
+				Type:     schema.TypeSet,
+				Optional: true,
+				Computed: true,
+				Set:      literalTypeHashCodeForSets,
+				Elem: &schema.Schema{
+					Type: schema.TypeString,
+				},
+			},
+			"private_endpoint_ip": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Computed: true,
+				ForceNew: true,
+			},
+
+			// Computed
+			"endpoint_fqdn": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"private_endpoint_id": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"state": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"time_created": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+		},
+	}
+}
+
+func createDataSafeDataSafePrivateEndpoint(d *schema.ResourceData, m interface{}) error {
+	sync := &DataSafeDataSafePrivateEndpointResourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).dataSafeClient
+
+	return CreateResource(d, sync)
+}
+
+func readDataSafeDataSafePrivateEndpoint(d *schema.ResourceData, m interface{}) error {
+	sync := &DataSafeDataSafePrivateEndpointResourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).dataSafeClient
+
+	return ReadResource(sync)
+}
+
+func updateDataSafeDataSafePrivateEndpoint(d *schema.ResourceData, m interface{}) error {
+	sync := &DataSafeDataSafePrivateEndpointResourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).dataSafeClient
+
+	return UpdateResource(d, sync)
+}
+
+func deleteDataSafeDataSafePrivateEndpoint(d *schema.ResourceData, m interface{}) error {
+	sync := &DataSafeDataSafePrivateEndpointResourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).dataSafeClient
+	sync.DisableNotFoundRetries = true
+
+	return DeleteResource(d, sync)
+}
+
+type DataSafeDataSafePrivateEndpointResourceCrud struct {
+	BaseCrud
+	Client                 *oci_data_safe.DataSafeClient
+	Res                    *oci_data_safe.DataSafePrivateEndpoint
+	DisableNotFoundRetries bool
+}
+
+func (s *DataSafeDataSafePrivateEndpointResourceCrud) ID() string {
+	return *s.Res.Id
+}
+
+func (s *DataSafeDataSafePrivateEndpointResourceCrud) CreatedPending() []string {
+	return []string{
+		string(oci_data_safe.LifecycleStateCreating),
+	}
+}
+
+func (s *DataSafeDataSafePrivateEndpointResourceCrud) CreatedTarget() []string {
+	return []string{
+		string(oci_data_safe.LifecycleStateActive),
+	}
+}
+
+func (s *DataSafeDataSafePrivateEndpointResourceCrud) DeletedPending() []string {
+	return []string{
+		string(oci_data_safe.LifecycleStateDeleting),
+	}
+}
+
+func (s *DataSafeDataSafePrivateEndpointResourceCrud) DeletedTarget() []string {
+	return []string{
+		string(oci_data_safe.LifecycleStateDeleted),
+	}
+}
+
+func (s *DataSafeDataSafePrivateEndpointResourceCrud) Create() error {
+	request := oci_data_safe.CreateDataSafePrivateEndpointRequest{}
+
+	if compartmentId, ok := s.D.GetOkExists("compartment_id"); ok {
+		tmp := compartmentId.(string)
+		request.CompartmentId = &tmp
+	}
+
+	if definedTags, ok := s.D.GetOkExists("defined_tags"); ok {
+		convertedDefinedTags, err := mapToDefinedTags(definedTags.(map[string]interface{}))
+		if err != nil {
+			return err
+		}
+		request.DefinedTags = convertedDefinedTags
+	}
+
+	if description, ok := s.D.GetOkExists("description"); ok {
+		tmp := description.(string)
+		request.Description = &tmp
+	}
+
+	if displayName, ok := s.D.GetOkExists("display_name"); ok {
+		tmp := displayName.(string)
+		request.DisplayName = &tmp
+	}
+
+	if freeformTags, ok := s.D.GetOkExists("freeform_tags"); ok {
+		request.FreeformTags = objectMapToStringMap(freeformTags.(map[string]interface{}))
+	}
+
+	if nsgIds, ok := s.D.GetOkExists("nsg_ids"); ok {
+		set := nsgIds.(*schema.Set)
+		interfaces := set.List()
+		tmp := make([]string, len(interfaces))
+		for i := range interfaces {
+			if interfaces[i] != nil {
+				tmp[i] = interfaces[i].(string)
+			}
+		}
+		if len(tmp) != 0 || s.D.HasChange("nsg_ids") {
+			request.NsgIds = tmp
+		}
+	}
+
+	if privateEndpointIp, ok := s.D.GetOkExists("private_endpoint_ip"); ok {
+		tmp := privateEndpointIp.(string)
+		request.PrivateEndpointIp = &tmp
+	}
+
+	if subnetId, ok := s.D.GetOkExists("subnet_id"); ok {
+		tmp := subnetId.(string)
+		request.SubnetId = &tmp
+	}
+
+	if vcnId, ok := s.D.GetOkExists("vcn_id"); ok {
+		tmp := vcnId.(string)
+		request.VcnId = &tmp
+	}
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(s.DisableNotFoundRetries, "data_safe")
+
+	response, err := s.Client.CreateDataSafePrivateEndpoint(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	workId := response.OpcWorkRequestId
+	return s.getDataSafePrivateEndpointFromWorkRequest(workId, getRetryPolicy(s.DisableNotFoundRetries, "data_safe"), oci_data_safe.WorkRequestResourceActionTypeCreated, s.D.Timeout(schema.TimeoutCreate))
+}
+
+func (s *DataSafeDataSafePrivateEndpointResourceCrud) getDataSafePrivateEndpointFromWorkRequest(workId *string, retryPolicy *oci_common.RetryPolicy,
+	actionTypeEnum oci_data_safe.WorkRequestResourceActionTypeEnum, timeout time.Duration) error {
+
+	// Wait until it finishes
+	dataSafePrivateEndpointId, err := dataSafePrivateEndpointWaitForWorkRequest(workId, "datasafeprivateendpoints",
+		actionTypeEnum, timeout, s.DisableNotFoundRetries, s.Client)
+
+	if err != nil {
+		log.Printf("[DEBUG] operation failed: %v for identifier: %v\n", workId, dataSafePrivateEndpointId)
+		return err
+	}
+	s.D.SetId(*dataSafePrivateEndpointId)
+
+	return s.Get()
+}
+
+func dataSafePrivateEndpointWorkRequestShouldRetryFunc(timeout time.Duration) func(response oci_common.OCIOperationResponse) bool {
+	startTime := time.Now()
+	stopTime := startTime.Add(timeout)
+	return func(response oci_common.OCIOperationResponse) bool {
+
+		// Stop after timeout has elapsed
+		if time.Now().After(stopTime) {
+			return false
+		}
+
+		// Make sure we stop on default rules
+		if shouldRetry(response, false, "data_safe", startTime) {
+			return true
+		}
+
+		// Only stop if the time Finished is set
+		if workRequestResponse, ok := response.Response.(oci_data_safe.GetWorkRequestResponse); ok {
+			return workRequestResponse.TimeFinished == nil
+		}
+		return false
+	}
+}
+
+func dataSafePrivateEndpointWaitForWorkRequest(wId *string, entityType string, action oci_data_safe.WorkRequestResourceActionTypeEnum,
+	timeout time.Duration, disableFoundRetries bool, client *oci_data_safe.DataSafeClient) (*string, error) {
+	retryPolicy := getRetryPolicy(disableFoundRetries, "data_safe")
+	retryPolicy.ShouldRetryOperation = dataSafePrivateEndpointWorkRequestShouldRetryFunc(timeout)
+
+	response := oci_data_safe.GetWorkRequestResponse{}
+	stateConf := &resource.StateChangeConf{
+		Pending: []string{
+			string(oci_data_safe.WorkRequestStatusInProgress),
+			string(oci_data_safe.WorkRequestStatusAccepted),
+		},
+		Target: []string{
+			string(oci_data_safe.WorkRequestStatusSucceeded),
+			string(oci_data_safe.WorkRequestStatusFailed),
+		},
+		Refresh: func() (interface{}, string, error) {
+			var err error
+			response, err = client.GetWorkRequest(context.Background(),
+				oci_data_safe.GetWorkRequestRequest{
+					WorkRequestId: wId,
+					RequestMetadata: oci_common.RequestMetadata{
+						RetryPolicy: retryPolicy,
+					},
+				})
+			wr := &response.WorkRequest
+			return wr, string(wr.Status), err
+		},
+		Timeout: timeout,
+	}
+	if _, e := stateConf.WaitForState(); e != nil {
+		return nil, e
+	}
+
+	var identifier *string
+	// The work request response contains an array of objects that finished the operation
+	for _, res := range response.Resources {
+		if strings.Contains(strings.ToLower(*res.EntityType), entityType) {
+			if res.ActionType == action {
+				identifier = res.Identifier
+				break
+			}
+		}
+	}
+	// The workrequest may have failed, check for errors if identifier is not found or work failed
+	if identifier == nil || response.Status == oci_data_safe.WorkRequestStatusFailed {
+		return nil, getErrorFromDataSafePrivateEndpointWorkRequest(client, wId, retryPolicy, entityType, action)
+	}
+
+	return identifier, nil
+}
+
+func getErrorFromDataSafePrivateEndpointWorkRequest(client *oci_data_safe.DataSafeClient, wId *string, retryPolicy *oci_common.RetryPolicy, entityType string, action oci_data_safe.WorkRequestResourceActionTypeEnum) error {
+	response, err := client.ListWorkRequestErrors(context.Background(),
+		oci_data_safe.ListWorkRequestErrorsRequest{
+			WorkRequestId: wId,
+			RequestMetadata: oci_common.RequestMetadata{
+				RetryPolicy: retryPolicy,
+			},
+		})
+	if err != nil {
+		return err
+	}
+	allErrs := make([]string, 0)
+	for _, wrkErr := range response.Items {
+		allErrs = append(allErrs, *wrkErr.Message)
+	}
+	errorMessage := strings.Join(allErrs, "\n")
+	workRequestErr := fmt.Errorf("work request did not succeed, workId: %s, entity: %s, action: %s. Message: %s", *wId, entityType, action, errorMessage)
+	return workRequestErr
+}
+
+func (s *DataSafeDataSafePrivateEndpointResourceCrud) Get() error {
+	request := oci_data_safe.GetDataSafePrivateEndpointRequest{}
+
+	tmp := s.D.Id()
+	request.DataSafePrivateEndpointId = &tmp
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(s.DisableNotFoundRetries, "data_safe")
+
+	response, err := s.Client.GetDataSafePrivateEndpoint(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	s.Res = &response.DataSafePrivateEndpoint
+	return nil
+}
+
+func (s *DataSafeDataSafePrivateEndpointResourceCrud) Update() error {
+	if compartment, ok := s.D.GetOkExists("compartment_id"); ok && s.D.HasChange("compartment_id") {
+		oldRaw, newRaw := s.D.GetChange("compartment_id")
+		if newRaw != "" && oldRaw != "" {
+			err := s.updateCompartment(compartment)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	request := oci_data_safe.UpdateDataSafePrivateEndpointRequest{}
+
+	tmp := s.D.Id()
+	request.DataSafePrivateEndpointId = &tmp
+
+	if definedTags, ok := s.D.GetOkExists("defined_tags"); ok {
+		convertedDefinedTags, err := mapToDefinedTags(definedTags.(map[string]interface{}))
+		if err != nil {
+			return err
+		}
+		request.DefinedTags = convertedDefinedTags
+	}
+
+	if description, ok := s.D.GetOkExists("description"); ok {
+		tmp := description.(string)
+		request.Description = &tmp
+	}
+
+	if displayName, ok := s.D.GetOkExists("display_name"); ok {
+		tmp := displayName.(string)
+		request.DisplayName = &tmp
+	}
+
+	if freeformTags, ok := s.D.GetOkExists("freeform_tags"); ok {
+		request.FreeformTags = objectMapToStringMap(freeformTags.(map[string]interface{}))
+	}
+
+	if nsgIds, ok := s.D.GetOkExists("nsg_ids"); ok {
+		set := nsgIds.(*schema.Set)
+		interfaces := set.List()
+		tmp := make([]string, len(interfaces))
+		for i := range interfaces {
+			if interfaces[i] != nil {
+				tmp[i] = interfaces[i].(string)
+			}
+		}
+		if len(tmp) != 0 || s.D.HasChange("nsg_ids") {
+			request.NsgIds = tmp
+		}
+	}
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(s.DisableNotFoundRetries, "data_safe")
+
+	response, err := s.Client.UpdateDataSafePrivateEndpoint(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	workId := response.OpcWorkRequestId
+	return s.getDataSafePrivateEndpointFromWorkRequest(workId, getRetryPolicy(s.DisableNotFoundRetries, "data_safe"), oci_data_safe.WorkRequestResourceActionTypeUpdated, s.D.Timeout(schema.TimeoutUpdate))
+}
+
+func (s *DataSafeDataSafePrivateEndpointResourceCrud) Delete() error {
+	request := oci_data_safe.DeleteDataSafePrivateEndpointRequest{}
+
+	tmp := s.D.Id()
+	request.DataSafePrivateEndpointId = &tmp
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(s.DisableNotFoundRetries, "data_safe")
+
+	response, err := s.Client.DeleteDataSafePrivateEndpoint(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	workId := response.OpcWorkRequestId
+	// Wait until it finishes
+	_, delWorkRequestErr := dataSafePrivateEndpointWaitForWorkRequest(workId, "datasafeprivateendpoints",
+		oci_data_safe.WorkRequestResourceActionTypeDeleted, s.D.Timeout(schema.TimeoutDelete), s.DisableNotFoundRetries, s.Client)
+	return delWorkRequestErr
+}
+
+func (s *DataSafeDataSafePrivateEndpointResourceCrud) SetData() error {
+	if s.Res.CompartmentId != nil {
+		s.D.Set("compartment_id", *s.Res.CompartmentId)
+	}
+
+	if s.Res.DefinedTags != nil {
+		s.D.Set("defined_tags", definedTagsToMap(s.Res.DefinedTags))
+	}
+
+	if s.Res.Description != nil {
+		s.D.Set("description", *s.Res.Description)
+	}
+
+	if s.Res.DisplayName != nil {
+		s.D.Set("display_name", *s.Res.DisplayName)
+	}
+
+	if s.Res.EndpointFqdn != nil {
+		s.D.Set("endpoint_fqdn", *s.Res.EndpointFqdn)
+	}
+
+	s.D.Set("freeform_tags", s.Res.FreeformTags)
+
+	nsgIds := []interface{}{}
+	for _, item := range s.Res.NsgIds {
+		nsgIds = append(nsgIds, item)
+	}
+	s.D.Set("nsg_ids", schema.NewSet(literalTypeHashCodeForSets, nsgIds))
+
+	if s.Res.PrivateEndpointId != nil {
+		s.D.Set("private_endpoint_id", *s.Res.PrivateEndpointId)
+	}
+
+	if s.Res.PrivateEndpointIp != nil {
+		s.D.Set("private_endpoint_ip", *s.Res.PrivateEndpointIp)
+	}
+
+	s.D.Set("state", s.Res.LifecycleState)
+
+	if s.Res.SubnetId != nil {
+		s.D.Set("subnet_id", *s.Res.SubnetId)
+	}
+
+	if s.Res.TimeCreated != nil {
+		s.D.Set("time_created", s.Res.TimeCreated.String())
+	}
+
+	if s.Res.VcnId != nil {
+		s.D.Set("vcn_id", *s.Res.VcnId)
+	}
+
+	return nil
+}
+
+func (s *DataSafeDataSafePrivateEndpointResourceCrud) updateCompartment(compartment interface{}) error {
+	changeCompartmentRequest := oci_data_safe.ChangeDataSafePrivateEndpointCompartmentRequest{}
+
+	compartmentTmp := compartment.(string)
+	changeCompartmentRequest.CompartmentId = &compartmentTmp
+
+	idTmp := s.D.Id()
+	changeCompartmentRequest.DataSafePrivateEndpointId = &idTmp
+
+	changeCompartmentRequest.RequestMetadata.RetryPolicy = getRetryPolicy(s.DisableNotFoundRetries, "data_safe")
+
+	_, err := s.Client.ChangeDataSafePrivateEndpointCompartment(context.Background(), changeCompartmentRequest)
+	if err != nil {
+		return err
+	}
+	return nil
+}