+++ oci/datacatalog_connection_resource.go
@@ -402,7 +402,7 @@ func (s *DatacatalogConnectionResourceCrud) SetData() error {
 	}
 
 	if s.Res.TimeStatusUpdated != nil {
-		s.D.Set("time_status_updated", *s.Res.TimeStatusUpdated)
+		s.D.Set("time_status_updated", s.Res.TimeStatusUpdated.String())
 	}
 
 	if s.Res.TimeUpdated != nil {