+++ oci/kms_vault_data_source.go
@@ -95,7 +95,7 @@ func (s *KmsVaultDataSourceCrud) SetData() error {
 	}
 
 	if s.Res.TimeOfDeletion != nil {
-		s.D.Set("time_of_deletion", *s.Res.TimeOfDeletion)
+		s.D.Set("time_of_deletion", s.Res.TimeOfDeletion.String())
 	}
 
 	s.D.Set("vault_type", s.Res.VaultType)